tests basicos para dataSourceSQL pasan

+++ proyecto_test_flutter_1/lib/Data/sql_data_source.dart
@@ -11,6 +11,9 @@ class SQLdataSource implements DataSource {
       dbPath,
       version: 1,
       onCreate: _createDB,
+      onOpen: (db) async {
+        await db.execute("PRAGMA foreign_keys = ON;");
+      },
     );
   }
 
@@ -21,15 +24,14 @@ class SQLdataSource implements DataSource {
         apellidos TEXT,
         email TEXT,
         sobreMi TEXT, 
+        avatarIndex INTEGER,
         UNIQUE (nombres, apellidos)
 
       )''');
     await db.execute('''CREATE TABLE relaciones(
-        persona1 TEXT,
-        persona2 TEXT,
+        persona1 TEXT REFERENCES personas(id) ON DELETE CASCADE,
+        persona2 TEXT REFERENCES personas(id) ON DELETE CASCADE,
         PRIMARY KEY(persona1, persona2),
-        FOREIGN KEY (persona1) REFERENCES personas(id) ON DELETE CASCADE,
-        FOREIGN KEY (persona2) REFERENCES personas(id) ON DELETE CASCADE,
         CHECK (persona1 != persona2)
       )''');
   }
@@ -37,11 +39,13 @@ class SQLdataSource implements DataSource {
   @override
   addPerson(Persona persona) async {
     sql.Database db = await _getDB();
-    await db.insert('personas', persona.toMap(),
+    int result = await db.insert('personas', persona.toMap(),
         conflictAlgorithm: sql.ConflictAlgorithm.ignore);
-    for (String id in persona.relaciones) {
-      await db.insert('relaciones', {'persona1': persona.id, 'persona2': id},
-          conflictAlgorithm: sql.ConflictAlgorithm.ignore);
+    if (result != 0) {
+      for (String id in persona.relaciones) {
+        await db.insert('relaciones', {'persona1': persona.id, 'persona2': id},
+            conflictAlgorithm: sql.ConflictAlgorithm.ignore);
+      }
     }
   }
 
@@ -56,23 +60,22 @@ class SQLdataSource implements DataSource {
   editPerson(Persona persona) async {
     sql.Database db = await _getDB();
 
-    await db.update('persona', persona.toMap(),
+    int result = await db.update('personas', persona.toMap(),
         where: 'id = ?',
         whereArgs: [persona.id],
         conflictAlgorithm: sql.ConflictAlgorithm.ignore);
 
-    await db.delete('relaciones',
-        where:
-            '(persona1 = ? AND persona2 NOT IN ?) OR (persona2 = ? AND persona1 NOT IN ?)',
-        whereArgs: [
-          persona.id,
-          persona.relaciones,
-          persona.id,
-          persona.relaciones
-        ]);
-    for (String id in persona.relaciones) {
-      await db.insert('relaciones', {'persona1': persona.id, 'persona2': id},
-          conflictAlgorithm: sql.ConflictAlgorithm.ignore);
+    if (result != 0) {
+      await db.delete('relaciones',
+          where: '(persona1 = ? OR persona2 = ?)',
+          whereArgs: [
+            persona.id,
+            persona.id,
+          ]);
+      for (String id in persona.relaciones) {
+        await db.insert('relaciones', {'persona1': persona.id, 'persona2': id},
+            conflictAlgorithm: sql.ConflictAlgorithm.ignore);
+      }
     }
   }
 
@@ -88,8 +91,7 @@ class SQLdataSource implements DataSource {
       return null;
     }
 
-    Map<String, String> personaJSON =
-        personasQueryResult[0] as Map<String, String>;
+    final personaJSON = personasQueryResult[0];
 
     String personaID = personaJSON['id'] as String;
 
@@ -115,5 +117,64 @@ class SQLdataSource implements DataSource {
     sql.Database db = await _getDB();
 
     db.delete('personas');
+  }
+
+  @override
+  Future<List<Persona>> retrieveAll() async {
+    sql.Database db = await _getDB();
+
+    final personasQueryResult = await db.query('personas');
+    final relacionesQueryResult = await db.query('relaciones');
+
+    List<Persona> personas = [];
+
+    for (final personaJSON in personasQueryResult) {
+      String personaID = personaJSON['id'] as String;
+      List<String> relaciones = [];
+      for (final relacionJSON in relacionesQueryResult) {
+        if (relacionJSON['persona1'] == personaID &&
+            !relaciones.contains(relacionJSON['persona2'])) {
+          relaciones.add(relacionJSON['persona2'] as String);
+        } else if (relacionJSON['persona2'] == personaID &&
+            !relaciones.contains(relacionJSON['persona1'])) {
+          relaciones.add(relacionJSON['persona1'] as String);
+        }
+      }
+      personas.add(Persona.fromJSON(personaJSON, relaciones));
+    }
+
+    return personas;
+  }
+
+  @override
+  Future<Persona?> retrievePersonById(String id) async {
+    sql.Database db = await _getDB();
+
+    final personasQueryResult =
+        await db.query('personas', where: 'id = ?', whereArgs: [id]);
+
+    if (personasQueryResult.isEmpty) {
+      return null;
+    }
+
+    final personaJSON = personasQueryResult[0];
+
+    String personaID = id;
+
+    final relacionesQueryResult = await db.query('relaciones',
+        where: 'persona1 = ? OR persona2 = ?',
+        whereArgs: [personaID, personaID]);
+
+    List<String> relaciones = [];
+    for (final relacionJSON in relacionesQueryResult) {
+      if (relacionJSON['persona1'] != personaID &&
+          !relaciones.contains(relacionJSON['persona1'])) {
+        relaciones.add(relacionJSON['persona1'] as String);
+      } else if (relacionJSON['persona2'] != personaID &&
+          !relaciones.contains(relacionJSON['persona2'])) {
+        relaciones.add(relacionJSON['persona2'] as String);
+      }
+    }
+    return Persona.fromJSON(personaJSON, relaciones);
   }
 }
